Split data cleaning into 2 files
Raw Data Standardization is a file that converts the raw data we recieved into a simulated raw dataset that should be extracted from the NUHS system once proper data entry protocols are put in place.

Data Cleaning refers to the general cleaning that we will be doing to the raw data files with proper data entry protocol.

In best case scenario, Raw Data Standardization should only ever be called once, as we will be educating our client on proper data entry.